更新代码

+++ soc-spring-boot-api/src/main/java/com/treeyh/example/springboot/api/enums/SexEnum.java
@@ -8,20 +8,20 @@ package com.treeyh.example.springboot.api.enums;
  */
 public enum SexEnum {
 
-    UNKNOWN((byte)0, "未知"),
-    MAN((byte)1, "男"),
-    WOMAN((byte)2, "女");
+    UNKNOWN(0, "未知"),
+    MAN(1, "男"),
+    WOMAN(2, "女");
 
-    private Byte code;
+    private Integer code;
     private String desc;
 
-    SexEnum(Byte code, String desc) {
+    SexEnum(Integer code, String desc) {
         this.code = code;
         this.desc = desc;
     }
 
 
-    public Byte getCode() {
+    public Integer getCode() {
         return code;
     }
